Add a slightly better prompt to the REPL
